+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Add IsParentSampled to ITransactionContext ([#1128](https://github.com/getsentry/sentry-dotnet/pull/1128)
 - Avoid warn in global mode ([#1132](https://github.com/getsentry/sentry-dotnet/pull/1132))
+- Fix `ParentSampledId` being reset on `Transaction` ([#1130](https://github.com/getsentry/sentry-dotnet/pull/1130))
 
 ## 3.8.1